# gen/openapi/rpc/bidderapi/v1/bidderapi.swagger.yaml
swagger: "2.0"
info:
  title: Bidder API
  version: 1.0.0-alpha
  license:
    name: Business Source License 1.1
    url: https://github.com/primevprotocol/mev-commit/blob/main/LICENSE
consumes:
  - application/json
produces:
  - application/json
paths:
  /v1/bidder/bid:
    post:
      summary: SendBid
      description: Send a bid to the bidder mev-commit node.
      operationId: Bidder_SendBid
      responses:
        "200":
          description: A successful response.(streaming responses)
          schema:
            type: object
            properties:
              result:
                $ref: '#/definitions/v1Commitment'
              error:
                $ref: '#/definitions/googlerpcStatus'
            title: Stream result of v1Commitment
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/googlerpcStatus'
      parameters:
        - name: body
          description: Unsigned bid message from bidders to the bidder mev-commit node.
          in: body
          required: true
          schema:
            $ref: '#/definitions/rpcbidderapiv1Bid'
  /v1/bidder/get_min_stake:
    get:
      summary: GetMinStake
      description: GetMinStake is called by the bidder to get the minimum stake required to be in the bidder registry.
      operationId: Bidder_GetMinStake
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/rpcbidderapiv1StakeResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/googlerpcStatus'
  /v1/bidder/get_stake:
    get:
      summary: GetStake
      description: GetStake is called by the bidder to get its stake in the bidder registry.
      operationId: Bidder_GetStake
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/rpcbidderapiv1StakeResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/googlerpcStatus'
  /v1/bidder/register_stake/{amount}:
    post:
      summary: RegisterStake
      description: RegisterStake is called by the bidder to register its stake in the bidder registry.
      operationId: Bidder_RegisterStake
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/rpcbidderapiv1StakeResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/googlerpcStatus'
      parameters:
        - name: amount
          in: path
          required: true
          type: string
definitions:
  googlerpcStatus:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      details:
        type: array
        items:
          type: object
          $ref: '#/definitions/protobufAny'
  protobufAny:
    type: object
    properties:
      '@type':
        type: string
    additionalProperties: {}
  rpcbidderapiv1Bid:
    type: object
    example:
      amount: 1e+18
      blockNumber: 123456
      txHash: 91a89B633194c0D86C539A1A5B14DCCacfD47094
    properties:
      txHash:
        type: string
        description: Hex string encoding of the hash of the transaction that the bidder wants to include in the block.
        pattern: '[a-fA-F0-9]{40}'
      amount:
        type: string
        format: int64
        description: Amount of ETH that the bidder is willing to pay to the provider for including the transaction in the block.
      blockNumber:
        type: string
        format: int64
        description: Max block number that the bidder wants to include the transaction in.
    description: Unsigned bid message from bidders to the bidder mev-commit node.
    title: Bid message
    required:
      - txHash
      - amount
      - blockNumber
  rpcbidderapiv1StakeResponse:
    type: object
    example:
      amount: "1000000000000000000"
    properties:
      amount:
        type: string
    description: Get staked amount for bidder in the bidder provider_registry.
    title: Stake response
  v1Commitment:
    type: object
    properties:
      txHash:
        type: string
        description: Hex string encoding of the hash of the transaction that the bidder wants to include in the block.
        pattern: '[a-fA-F0-9]{40}'
      bidAmount:
        type: string
        format: int64
        description: Amount of ETH that the bidder has agreed to pay to the provider for including the transaction in the block.
      blockNumber:
        type: string
        format: int64
        description: Max block number that the bidder wants to include the transaction in.
      receivedBidDigest:
        type: string
        description: Hex string encoding of digest of the bid message signed by the bidder.
      receivedBidSignature:
        type: string
        description: Hex string encoding of signature of the bidder that sent this bid.
      commitmentDigest:
        type: string
        description: Hex string encoding of digest of the commitment.
      commitmentSignature:
        type: string
        description: Hex string encoding of signature of the commitment signed by the provider confirming this transaction.